refactor: DAO를 인터페이스로 선언하여 교체 가능하게 함 - 기존 DAO들에 InMemory 접두어를 붙인다.
